Work around IE11 speed reset on resume bug (#1)
IE11 sets videotag.playbackRate = 1 on resume, also triggering the
speed event.
Thanks to the beforeresume event we get our foot in the door just in
time, even if a speed change is done in paused state.

We do not bother about the speed change _display_ on resume in browsers
affected because this is a browser bug.